Stop reusing GameServer objects, cache InetAddresses instead
Based on my debugging, there seems to be an issue with the query library causing server player retrieval to enter an invalid state for a given GameServer object sometimes. Rather than trying to track this down in steam-condenser, we can just return to the old bot behaviour of always remaking GameServer objects, since caching InetAddresses was initially the only optimization needed anyways.